solving issue

+++ Controllers/Auth/UserAuthController.cs
@@ -30,7 +30,7 @@ public class UserAuthController :ControllerBase
 
     // GET: api/Users/5
     [HttpGet("{id}")]
-    public async Task<ActionResult<User>> GetUser(long id)
+    public async Task<ActionResult<User>> GetUser(int id)
     {
         var user = await _context.Users.FindAsync(id);
